Initial query commit
Add support for new query API.

The riak_kv_query module allows a query to be constructed with validation checks.

The riak_kv_query_server is a gen_Server for managing a query.  The coverage_fsm behaviour is not used, to aid future migration from the deprecated fsm.

There is a riak_kv_query_buffer added to replace the use of riak_kv_fold_buffer - which assumes the accumulator is a list.