Add raw_keys support and improve performance
The de-duplication of keys using lists:umerge/2 does not scale effectively,  Now, beyond a limit a temporary ets table is used instead for de-duplication.

Still, the performance was slower than current 2i for large key sets > 10K.  This is as riak_kv_query_server is the bottleneck - it cannot ACK batches of keys whilst it is hanlding previous batches.  In contrast the existing 2i simply forwards batches immediately - so that any delay processing is on the riak_client process queue, not on the query coordinator.

To help with this performance issue, the default batch size is increased, and also jittered to help avoid all vnode workers sending concurrently.

The use of an ets table requires every key to be wrapped in a tuple.  This is done at source (in the buffer), to avoid a further delay on the query server running a lists:map/2 function.  The JSON encoder will handle the tuple wrapping of the key (although it could be efficiently removed in the ets case using a select statement rather than tab2list.