Reap on all available primaries
Defer the reap only for unavailable primaries.

Previously the reap was deferred on all primaries if some were not available.  This caused issues with multi-cluster environments - where a set of reaps would occur on one cluster, but not on another (due to a node being unavailable).  This caused full-sync reconciliation issues and the resurrection of tombstones.