Convert riak object metadata into a map
The map module is mroe performant that the dict module, in part as dict is coupled to an old (deprecated) hash algorithm.

All objects are stored in binary form - which is independent of the data type.  This code will still handle objects converting metadata to dict.  However, enabling the switch to version 1 will require all connected nodes to support metadata as map.

A capability could be used, but there is history of capability registrations taking a long time to remove - so the shortcut of making this an operator responsibility has been made, especially with a major version change pending.