Two additional summary pages for 3.4
Further summary pages to describe best practice for Riak 3.4.

Where possible, reference to specific commands is avoided .  Any external product is only referenced when it is known to have been tested.

The aim is to prefer accuracy and maintainability of documentation, at the expense completeness.  Brevity is preferred to providing detailed instructions to inexpert users.

It is hoped that seven pages may provide a guide to a legacy-free description of Riak 3.4:
- InitialDesignDecisions (https://github.com/OpenRiak/riak_kv/pull/61);
- InstallAndStartGuide;
- BuildAndScaleClusterGuide;
- ObjectAPI (TODO);
- QueryAPI (https://github.com/OpenRiak/riak_kv/pull/51);
- NextGenREPL - GettingStarted;
- OperationsGuide (TODO).