Fix "a number of" issues
Saying "a number of" is flagged by AI as being vague.  Sometimes vagueness was deliberate, to avoid counting and listing everything in a way that would be brittle to future changes.

However, by rule, I think "a number of" can be considered a sign of lazy writing.  So each instance is addressed.